tests/requirements.txt
# Test requirements for Auto Cover integration
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.13.0
freezegun>=1.2.0
syrupy>=4.0.0